compose-sample-4/flask-mysql-in-docker/requirements.txt
Flask==1.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.11
pytest==5.4.1
pymysql==0.9.3

